Added three Files
ps1a is the initial program that calculates a simple predicted time of down payment using yearly salary and percentage saved. ps1b is similar to ps1a but includes semi-annual raises. ps1c3 only requires an input of a yearly salary but has a time limit of 36 months, and will output an optimal percentage to save in those 36 months and will return the number of guesses it used to come to that conclusion, or if it is not possible to save enough money will tell the user just that.